Log errors in CoapClient.tryToConnect()
I'm trying to use a library which uses this library (node-tradfri-client), where I've managed to find out that the attempts to establish a connection fail in this method. This commits adds a debug call to attempt to give the user some context as to why the attempt to establish a connection fails.